Brick-34-fix: 디자인과 다른 부분 수정 & gray, white, black, error 컬러 tailwind config에 지정
