+++ app/(auth)/signin/page.tsx
@@ -1,0 +1,7 @@
+import { SignInForm } from '@/components/index'
+
+function SignInPage() {
+  return <SignInForm />
+}
+
+export default SignInPage
