+++ app/(auth)/signup/email/page.tsx
@@ -1,0 +1,7 @@
+import { SignUpEmailForm } from '@/components/index'
+
+function SignUpEmailPage() {
+  return <SignUpEmailForm />
+}
+
+export default SignUpEmailPage
