+++ app/(auth)/signup/page.tsx
@@ -1,0 +1,7 @@
+import { SignUpStartForm } from '@/components/index'
+
+function SignUpPage() {
+  return <SignUpStartForm />
+}
+
+export default SignUpPage
