--- app/signin/page.tsx
@@ -1,11 +1,0 @@
-import SignInForm from '@/components/SignInForm'
-
-function SignInPage() {
-  return (
-    <div>
-      <SignInForm />
-    </div>
-  )
-}
-
-export default SignInPage
