--- app/signup/email/page.tsx
@@ -1,12 +1,0 @@
-import SignUpEmail from '@/components/SignUpEmail'
-import React from 'react'
-
-function SignUpEmailPage() {
-  return (
-    <div>
-      <SignUpEmail />
-    </div>
-  )
-}
-
-export default SignUpEmailPage
