--- app/signup/page.tsx
@@ -1,11 +1,0 @@
-import SignUpStart from '@/components/SignUpStart'
-
-function SignUpPage() {
-  return (
-    <div>
-      <SignUpStart />
-    </div>
-  )
-}
-
-export default SignUpPage
