+++ components/index.ts
@@ -7,8 +7,8 @@ export { default as Input } from './Input'
 export { default as NestingAvatar } from './NestingAvatar'
 export { default as PeopleCounter } from './PeopleCounter/PeopleCounter'
 export { default as SignInForm } from './SignInForm'
-export { default as SignUpStart } from './SignUpStart'
-export { default as SignUpEmail } from './SignUpEmail'
+export { default as SignUpStartForm } from './SignUpStartForm'
+export { default as SignUpEmailForm } from './SignUpEmailForm'
 export { default as Tab } from './Tab'
 export { default as CrewGatheringInfo } from './CrewGatheringInfo/CrewGatheringInfo'
 export { default as IconLabel } from './CrewGatheringInfo/IconLabel'
